requirements.txt
numpy~=2.1.1
opencv-python~=4.10.0.84
pillow~=10.4.0
pynput~=1.7.7